requirements.txt

streamlit
yt-dlp
whisper-openai
moviepy
